@@ -1,30 +1,25 @@
 {
 	"name": "web",
-	"version": "0.0.1",
+	"version": "2.0.0",
 	"private": true,
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
-		"preview": "vite preview",
-		"lint": "prettier --plugin-search-dir . --check .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"preview": "vite preview"
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/adapter-static": "^2.0.1",
-		"@sveltejs/kit": "^1.5.0",
-		"autoprefixer": "^10.4.13",
-		"mdsvex": "^0.10.6",
-		"postcss": "^8.4.21",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
-		"svelte": "^3.56.0",
-		"tailwindcss": "^3.2.7",
-		"vite": "^4.0.0"
+		"@sveltejs/adapter-static": "^2.0.3",
+		"@sveltejs/kit": "^1.27.4",
+		"autoprefixer": "^10.4.16",
+		"postcss": "^8.4.31",
+		"postcss-load-config": "^4.0.2",
+		"svelte": "^4.0.5",
+		"tailwindcss": "^3.3.5",
+		"vite": "^4.4.2"
 	},
 	"type": "module",
 	"dependencies": {
-		"svelte-preprocess": "^5.0.1",
-		"thumb-ui": "^0.0.17"
+		"svelte-preprocess": "^5.1.0"
 	}
 }
