@@ -1,25 +1,133 @@
 <script>
     import "/src/app.css";
 
-    import Analytics from "$lib/Analytics.svelte";
-    import ThemeToggle from "$lib/ThemeToggle.svelte";
-    import Link from "$lib/Link.svelte";
+    import {onMount} from "svelte";
+    import Sidebar from "$lib/components/Sidebar.svelte";
+    import Banner from "$lib/components/Banner.svelte";
+    import Navigation from "$lib/components/Navigation.svelte";
+    import Link from "$lib/components/Link.svelte";
+
+    const githubURL = import.meta.env.VITE_GITHUB_URL;
+    const linkedinURL = import.meta.env.VITE_LINKEDIN_URL;
+    const trivinoURL = import.meta.env.VITE_TRIVINO_URL;
+    const nooklyURL = import.meta.env.VITE_NOOKLY_URL;
+
+    let show;
+    let loading = true;
+    onMount(() => {
+        loading = false;
+    });
 </script>
 
-<Analytics/>
+<style>
+    button {
+        @apply
+            rounded-md bg-black
+            flex items-center justify-center
+            cursor-pointer
+            lg:hidden
+            hover:bg-stroke
+    }
+</style>
 
-<div class="min-h-full flex flex-col items-start pb-4">
-    <ThemeToggle class="h-8 w-8 cursor-pointer mt-4 ml-4 lg:mt-8 lg:ml-8" />
-    <div class="h-full flex flex-col lg:flex-row gap-12 px-4 lg:py-36 lg:px-72">
-        <nav class="flex flex-row lg:flex-col gap-8 pt-16 lg:pt-24">
-            <Link href="/">home</Link>
-            <Link href="/about">about</Link>
-            <Link href="/blog">blog</Link>
-            <Link href="/gallery">gallery</Link>
-        </nav>
 
-        <div class="h-full flex flex-col">
-            <slot/>
-        </div>
+<div class="relative h-screen flex flex-row">
+    <Sidebar bind:show={show}>
+        <Banner border={false}>
+            <div class="w-full h-full flex flex-row gap-4 items-center pl-6">
+                <button on:click={() => show = !show} aria-expanded={show} aria-haspopup="true" aria-label="Nav toggle">
+                    <svg width="12" height="12" viewBox="0 0 20 18" fill="white" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M2 1L18 17" stroke="white" stroke-width="2" stroke-linecap="round"/>
+                        <path d="M2 17L18 1" stroke="white" stroke-width="2" stroke-linecap="round"/>
+                    </svg>
+                </button>
+                <span class="text-xs font-semibold">Matt Collecutt</span>
+            </div>
+        </Banner>
+        <Navigation>
+            <Link href="/">
+                <div slot="icon" >
+                    <svg width="12" height="12" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M15.9944 7.98438C15.9944 8.54688 15.5778 8.9875 15.1056 8.9875H14.2167L14.2361 13.9937C14.2361 14.0781 14.2306 14.1625 14.2222 14.2469V14.75C14.2222 15.4406 13.725 16 13.1111 16H12.6667C12.6361 16 12.6056 16 12.575 15.9969C12.5361 16 12.4972 16 12.4583 16H11.5556H10.8889C10.275 16 9.77778 15.4406 9.77778 14.75V14V12C9.77778 11.4469 9.38056 11 8.88889 11H7.11111C6.61944 11 6.22222 11.4469 6.22222 12V14V14.75C6.22222 15.4406 5.725 16 5.11111 16H4.44444H3.55833C3.51667 16 3.475 15.9969 3.43333 15.9937C3.4 15.9969 3.36667 16 3.33333 16H2.88889C2.275 16 1.77778 15.4406 1.77778 14.75V11.25C1.77778 11.2219 1.77778 11.1906 1.78056 11.1625V8.9875H0.888889C0.388889 8.9875 0 8.55 0 7.98438C0 7.70312 0.0833333 7.45312 0.277778 7.23438L7.4 0.25C7.59444 0.03125 7.81667 0 8.01111 0C8.20556 0 8.42778 0.0625 8.59444 0.21875L15.6889 7.23438C15.9111 7.45312 16.0222 7.70312 15.9944 7.98438Z" fill="white"/>
+                    </svg>
+                </div>
+                Home
+            </Link>
+            <Link href="/photos">
+                <div slot="icon">
+                    <svg width="12" height="10" viewBox="0 0 16 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M6.89375 2.7875L8.47188 2H14V5H10.4125C9.72813 4.52812 8.89688 4.25 8 4.25C7.10313 4.25 6.275 4.52812 5.5875 5H2V3H6C6.30937 3 6.61563 2.92812 6.89375 2.7875ZM0 3V12C0 13.1031 0.896875 14 2 14H14C15.1031 14 16 13.1031 16 12V2C16 0.896875 15.1031 0 14 0H8.47188C8.1625 0 7.85625 0.071875 7.57812 0.2125L6 1H5V0.5C5 0.225 4.775 0 4.5 0H2.5C2.225 0 2 0.225 2 0.5V1C0.896875 1 0 1.89688 0 3ZM5.25 8.5C5.25 7.77065 5.53973 7.07118 6.05546 6.55546C6.57118 6.03973 7.27065 5.75 8 5.75C8.72935 5.75 9.42882 6.03973 9.94454 6.55546C10.4603 7.07118 10.75 7.77065 10.75 8.5C10.75 9.22935 10.4603 9.92882 9.94454 10.4445C9.42882 10.9603 8.72935 11.25 8 11.25C7.27065 11.25 6.57118 10.9603 6.05546 10.4445C5.53973 9.92882 5.25 9.22935 5.25 8.5Z" fill="white"/>
+                    </svg>
+                </div>
+                Photos
+            </Link>
+            <Link href="/writing">
+                <div slot="icon">
+                    <svg width="12" height="12" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M11.5125 0.571875L9.7719 2.31562L13.6844 6.22813L15.425 4.4875C16.1094 3.80313 16.1094 2.69688 15.425 2.0125L13.9875 0.571875C13.3032 -0.1125 12.1969 -0.1125 11.5125 0.571875ZM9.00003 2.95625L8.71253 3.04375L4.2094 4.39375C3.58753 4.58125 3.09378 5.05625 2.88753 5.675L0.11878 13.9312C2.99215e-05 14.2844 0.0875299 14.6781 0.346905 14.9438L5.1469 10.1469C5.05315 9.95 5.00003 9.73125 5.00003 9.5C5.00003 8.67188 5.67191 8 6.50003 8C7.32816 8 8.00003 8.67188 8.00003 9.5C8.00003 10.3281 7.32816 11 6.50003 11C6.26878 11 6.05003 10.9469 5.85316 10.8531L1.05315 15.6531C1.3219 15.9125 1.71253 16.0031 2.06565 15.8813L10.325 13.1125C10.9407 12.9063 11.4188 12.4125 11.6063 11.7906L12.9563 7.2875L13.0438 7L9.00003 2.95625Z" fill="white"/>
+                    </svg>
+                </div>
+                Writing
+            </Link>
+        </Navigation>
+        <Navigation title="Projects">
+            <Link href={trivinoURL} external={true} target="_blank">
+                <div slot="icon" >
+                    <svg width="10" height="19" viewBox="0 0 50 92" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M41.6256 2H8.37441C8.37441 2 -13.9662 54.6172 25 54.6172C63.9662 54.6172 41.6256 2 41.6256 2Z" stroke="white" stroke-width="4"/>
+                        <line x1="26.0808" y1="56.7478" x2="26.0808" y2="85.911" stroke="white" stroke-width="4" stroke-linecap="square"/>
+                        <path d="M23.3128 87.6973L19.1564 89.3947M23.3128 87.6973V86.9139M23.3128 87.6973V88.546M19.1564 89.3947H18.7667M19.1564 89.3947H20.4553M23.3128 86.9139V86C18.6368 89.3947 15.7793 89.3947 15 89.3947H18.7667M23.3128 86.9139L18.7667 89.3947M23.3128 88.546V89.3947H20.4553M23.3128 88.546L20.4553 89.3947" stroke="white" stroke-width="2"/>
+                        <path d="M29 87.6973L33.1564 89.3947M29 87.6973V86.9139M29 87.6973V88.546M33.1564 89.3947H33.546M33.1564 89.3947H31.8575M29 86.9139V86C33.6759 89.3947 36.5334 89.3947 37.3127 89.3947H33.546M29 86.9139L33.546 89.3947M29 88.546V89.3947H31.8575M29 88.546L31.8575 89.3947" stroke="white" stroke-width="2"/>
+                        <rect x="7.45532" y="87.911" width="35.3293" height="4" rx="1" fill="white"/>
+                        <path d="M7.39599 33.1162C8.99456 43.2376 17.0198 49 26.0278 49C35.0358 49 43.788 39.9751 43.3996 34.0788C43.0113 28.1826 42.5539 26.6183 42.5539 26.6183C33.251 30.2282 21.9304 20 16.3365 20C9.57071 20.361 5.70455 22.4066 7.39599 33.1162Z" fill="#843737"/>
+                        <path d="M7.39599 33.1162C8.99456 43.2376 17.0198 49 26.0278 49C35.0358 49 43.788 39.9751 43.3996 34.0788C43.0113 28.1826 42.5539 26.6183 42.5539 26.6183C33.251 30.2282 21.9304 20 16.3365 20C9.57071 20.361 5.70455 22.4066 7.39599 33.1162Z" fill="#843737"/>
+                        <path d="M7.39599 33.1162C8.99456 43.2376 17.0198 49 26.0278 49C35.0358 49 43.788 39.9751 43.3996 34.0788C43.0113 28.1826 42.5539 26.6183 42.5539 26.6183C33.251 30.2282 21.9304 20 16.3365 20C9.57071 20.361 5.70455 22.4066 7.39599 33.1162Z" fill="#843737"/>
+                        <path d="M7.39599 33.1162C8.99456 43.2376 17.0198 49 26.0278 49C35.0358 49 43.788 39.9751 43.3996 34.0788C43.0113 28.1826 42.5539 26.6183 42.5539 26.6183C33.251 30.2282 21.9304 20 16.3365 20C9.57071 20.361 5.70455 22.4066 7.39599 33.1162Z" fill="#843737"/>
+                        <path d="M43.9741 33.5975C43.9741 43.8434 34.497 49 25.489 49C16.481 49 7.72883 43.5851 7.24555 34.1992C6.76228 24.8133 7.24555 26.7386 7.24555 26.7386C13.7697 36.9668 29.2304 19.3983 35.3049 20C42.0707 20.361 43.9741 22.0456 43.9741 33.5975Z" fill="#A93E3E"/>
+                        <path d="M43.9741 33.5975C43.9741 43.8434 34.497 49 25.489 49C16.481 49 7.72883 43.5851 7.24555 34.1992C6.76228 24.8133 7.24555 26.7386 7.24555 26.7386C13.7697 36.9668 29.2304 19.3983 35.3049 20C42.0707 20.361 43.9741 22.0456 43.9741 33.5975Z" fill="#A93E3E"/>
+                        <path d="M43.9741 33.5975C43.9741 43.8434 34.497 49 25.489 49C16.481 49 7.72883 43.5851 7.24555 34.1992C6.76228 24.8133 7.24555 26.7386 7.24555 26.7386C13.7697 36.9668 29.2304 19.3983 35.3049 20C42.0707 20.361 43.9741 22.0456 43.9741 33.5975Z" fill="#A93E3E"/>
+                        <path d="M43.9741 33.5975C43.9741 43.8434 34.497 49 25.489 49C16.481 49 7.72883 43.5851 7.24555 34.1992C6.76228 24.8133 7.24555 26.7386 7.24555 26.7386C13.7697 36.9668 29.2304 19.3983 35.3049 20C42.0707 20.361 43.9741 22.0456 43.9741 33.5975Z" fill="#A93E3E"/>
+                    </svg>
+
+                </div>
+                TriVino
+            </Link>
+            <Link href={nooklyURL} external={true} target="_blank">
+                <div slot="icon" >
+                    <svg width="11.9" height="12" viewBox="0 0 150 159" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M74.7104 43.1081C66.9884 43.1081 60.6178 45.1906 55.5984 49.3556C50.5791 53.3124 46.9112 58.6228 44.5946 65.2868C42.278 71.7426 41.1197 79.0314 41.1197 87.1532V138.44C41.1197 149.795 31.9147 159 20.5598 159C9.20496 159 0 149.795 0 138.44V81.8428C0 66.8487 2.7027 53.2082 8.10811 40.9214C13.7066 28.4263 22.0077 18.5344 33.0116 11.2456C44.0154 3.74853 57.9151 0 74.7104 0C91.5058 0 105.405 3.74853 116.409 11.2456C127.606 18.5344 136.004 28.3222 141.602 40.609C147.201 52.8959 150 66.6405 150 81.8428V138.44C150 149.795 140.795 159 129.44 159C118.085 159 108.88 149.795 108.88 138.44V87.1532C108.88 79.2397 107.625 71.9509 105.116 65.2868C102.606 58.6228 98.8417 53.3124 93.8224 49.3556C88.9961 45.1906 82.6255 43.1081 74.7104 43.1081Z" fill="url(#paint0_linear_320_1429)"/>
+                        <defs>
+                            <linearGradient id="paint0_linear_320_1429" x1="271.631" y1="-23.2191" x2="271.631" y2="149.634" gradientUnits="userSpaceOnUse">
+                                <stop stop-color="#FF411A"/>
+                                <stop offset="1" stop-color="#FF6C4D"/>
+                            </linearGradient>
+                        </defs>
+                    </svg>
+
+                </div>
+                Nookly
+            </Link>
+        </Navigation>
+        <Navigation title="Socials">
+            <Link href={githubURL} external={true} target="_blank">
+                <div slot="icon" >
+                    <svg width="12" height="12" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M5.35161 12.5613C5.35161 12.6258 5.27742 12.6774 5.18387 12.6774C5.07742 12.6871 5.00323 12.6355 5.00323 12.5613C5.00323 12.4968 5.07742 12.4452 5.17097 12.4452C5.26774 12.4355 5.35161 12.4871 5.35161 12.5613ZM4.34839 12.4161C4.32581 12.4806 4.39032 12.5548 4.4871 12.5742C4.57097 12.6065 4.66774 12.5742 4.6871 12.5097C4.70645 12.4452 4.64516 12.371 4.54839 12.3419C4.46452 12.3194 4.37097 12.3516 4.34839 12.4161ZM5.77419 12.3613C5.68064 12.3839 5.61613 12.4452 5.62581 12.5194C5.63548 12.5839 5.71935 12.6258 5.81613 12.6032C5.90968 12.5806 5.97419 12.5194 5.96452 12.4548C5.95484 12.3935 5.86774 12.3516 5.77419 12.3613ZM7.89677 0C3.42258 0 0 3.39677 0 7.87097C0 11.4484 2.25161 14.5097 5.46774 15.5871C5.88064 15.6613 6.02581 15.4065 6.02581 15.1968C6.02581 14.9968 6.01613 13.8935 6.01613 13.2161C6.01613 13.2161 3.75806 13.7 3.28387 12.2548C3.28387 12.2548 2.91613 11.3161 2.3871 11.0742C2.3871 11.0742 1.64839 10.5677 2.43871 10.5774C2.43871 10.5774 3.24194 10.6419 3.68387 11.4097C4.39032 12.6548 5.57419 12.2968 6.03548 12.0839C6.10968 11.5677 6.31935 11.2097 6.55161 10.9968C4.74839 10.7968 2.92903 10.5355 2.92903 7.43226C2.92903 6.54516 3.17419 6.1 3.69032 5.53226C3.60645 5.32258 3.33226 4.45806 3.77419 3.34194C4.44839 3.13226 6 4.2129 6 4.2129C6.64516 4.03226 7.33871 3.93871 8.02581 3.93871C8.7129 3.93871 9.40645 4.03226 10.0516 4.2129C10.0516 4.2129 11.6032 3.12903 12.2774 3.34194C12.7194 4.46129 12.4452 5.32258 12.3613 5.53226C12.8774 6.10323 13.1935 6.54839 13.1935 7.43226C13.1935 10.5452 11.2935 10.7935 9.49032 10.9968C9.7871 11.2516 10.0387 11.7355 10.0387 12.4935C10.0387 13.5806 10.029 14.9258 10.029 15.1903C10.029 15.4 10.1774 15.6548 10.5871 15.5806C13.8129 14.5097 16 11.4484 16 7.87097C16 3.39677 12.371 0 7.89677 0ZM3.13548 11.1258C3.09355 11.1581 3.10323 11.2323 3.15806 11.2935C3.20968 11.3452 3.28387 11.3677 3.32581 11.3258C3.36774 11.2935 3.35806 11.2194 3.30323 11.1581C3.25161 11.1065 3.17742 11.0839 3.13548 11.1258ZM2.7871 10.8645C2.76452 10.9065 2.79677 10.9581 2.86129 10.9903C2.9129 11.0226 2.97742 11.0129 3 10.9677C3.02258 10.9258 2.99032 10.8742 2.92581 10.8419C2.86129 10.8226 2.80968 10.8323 2.7871 10.8645ZM3.83226 12.0129C3.78065 12.0548 3.8 12.1516 3.87419 12.2129C3.94839 12.2871 4.04194 12.2968 4.08387 12.2452C4.12581 12.2032 4.10645 12.1065 4.04194 12.0452C3.97097 11.971 3.87419 11.9613 3.83226 12.0129ZM3.46452 11.5387C3.4129 11.571 3.4129 11.6548 3.46452 11.729C3.51613 11.8032 3.60323 11.8355 3.64516 11.8032C3.69677 11.7613 3.69677 11.6774 3.64516 11.6032C3.6 11.529 3.51613 11.4968 3.46452 11.5387Z" fill="white"/>
+                    </svg>
+                </div>
+                Github
+            </Link>
+            <Link href={linkedinURL} external={true} target="_blank">
+                <div slot="icon" >
+                    <svg width="12" height="12" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M14.8571 0H1.13929C0.510714 0 0 0.517857 0 1.15357V14.8464C0 15.4821 0.510714 16 1.13929 16H14.8571C15.4857 16 16 15.4821 16 14.8464V1.15357C16 0.517857 15.4857 0 14.8571 0ZM4.83571 13.7143H2.46429V6.07857H4.83929V13.7143H4.83571ZM3.65 5.03571C2.88929 5.03571 2.275 4.41786 2.275 3.66071C2.275 2.90357 2.88929 2.28571 3.65 2.28571C4.40714 2.28571 5.025 2.90357 5.025 3.66071C5.025 4.42143 4.41071 5.03571 3.65 5.03571ZM13.725 13.7143H11.3536V10C11.3536 9.11429 11.3357 7.975 10.1214 7.975C8.88571 7.975 8.69643 8.93929 8.69643 9.93572V13.7143H6.325V6.07857H8.6V7.12143H8.63214C8.95 6.52143 9.725 5.88929 10.8786 5.88929C13.2786 5.88929 13.725 7.47143 13.725 9.52857V13.7143Z" fill="white"/>
+                    </svg>
+                </div>
+                Linkedin
+            </Link>
+        </Navigation>
+    </Sidebar>
+    <div class="h-full w-full flex flex-col text-grey">
+        <slot/>
     </div>
 </div>
