@@ -1,30 +1,42 @@
 <script>
-    import { page } from "$lib/store.js";
-
-    $page = "/";
+    import Banner from "$lib/components/Banner.svelte";
 </script>
+
+<style>
+    a {
+        @apply
+            text-[#1F6FEB]
+            hover:underline
+        ;
+    }
+</style>
 
 <svelte:head>
     <title>Matt Collecutt</title>
+    <meta name="description" content="Home page" />
 </svelte:head>
 
-<h1>Matt Collecutt</h1>
 
-<article class="pt-12 flex flex-col gap-4">
-    <p>
-        Hi, I’m Matt! I’m passionate about a lot of things. Here you’ll find observations
-        I’ve made about different topics with a focus around software development.
-    </p>
+<Banner></Banner>
 
-    <p>
-        The main reason for starting this blog is to improve my written communication skills,
-        it’s always been somewhat of a struggle. I hope you’ll enjoy what I have to offer in
-        the form of blog posts and that you’ll leave having learnt something new.
-    </p>
+<div class="mx-auto w-full py-12 max-w-prose">
+    <article class="flex flex-col gap-5 text-sm font-light leading-6 tracking-wide max-w-prose">
+        <p>
+            Hi, I’m Matt! I'm a software developer, <a href="/photos">amateur photographer</a>,
+            <a href="/writing">writer</a>. Currently I'm working at
+            <a href="http://www.mustard.co.nz/" target="_blank">mustard technology</a>, a small development company
+            based in New Zealand.
+        </p>
 
-    <p>
-        I’m still exploring the direction I want to take this website, so the content may evolve
-        over time. I’ve love to hear your thoughts and feedback as I continue to develop the site,
-        so feel free to leave a comment or get in touch.
-    </p>
-</article>
+        <p>
+            Before Mustard, I worked at <a href="https://www.theaccessgroup.com/">Ezitracker</a>, where I started my
+            software development journey learning the ropes as an intern.
+        </p>
+
+        <p>
+            In my spare time I've been working on a few projects like <a href="https://mattcollecutt.co.nz">this</a>
+            website, <a href="https://trivino.xyz" target="_blank">TriVino</a> and
+            <a href="https://nookly.co.nz" target="_blank">Nookly</a>
+        </p>
+    </article>
+</div>
