@@ -1,25 +1,22 @@
 import adapter from '@sveltejs/adapter-static';
-import preprocess from "svelte-preprocess";
-import {mdsvex} from "mdsvex";
-
+import preprocess from 'svelte-preprocess'
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
+		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
+		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter({
 			pages: 'build',
 			assets: 'build',
 			fallback: 'index.html'
-		}),
-		prerender: {entries: ['/gallery/[slug]', '/blog/[slug]', '/', '/about', '/blog', '/gallery']},
-		inlineStyleThreshold: 1024
+		})
 	},
-	extensions: ['.svelte', '.md'],
 	preprocess: [
 		preprocess({
-			postcss: true
-		}),
-		mdsvex({ extensions: ['.md'] })
-	]
+			postcss: true,
+		})
+	],
 };
 
 export default config;
